Split geometry_models into a separate module
